Move the submission schedulers to miles/rollout/submission_scheduler.py
They serve two drivers in different modules and depend on nothing in
inference_rollout_common; trim comments to one line.
