+++ miles/rollout/inference_rollout/inference_rollout_common.py
@@ -120,66 +120,6 @@ async def generate_and_rm(
     return sample
 
 
-class GroupLevelSubmission:
-    """A submission slot frees only when the whole group task returns."""
-
-    sample_done_callback = None
-
-    def has_capacity(self, *, pending_groups: int, group_budget: int) -> bool:
-        return pending_groups < group_budget
-
-    def on_submit(self, groups: list[list[Sample]]) -> None:
-        pass
-
-    async def wait_for_progress(self, pendings: set[asyncio.Task]) -> tuple[set[asyncio.Task], set[asyncio.Task]]:
-        return await asyncio.wait(pendings, return_when=asyncio.FIRST_COMPLETED)
-
-
-class SampleBackfillSubmission:
-    """Each finished sample frees its own slot: a replacement group fits once
-    ``group_size`` samples complete, whichever groups they came from."""
-
-    def __init__(self, group_size: int):
-        self.group_size = group_size
-        self.samples_in_flight = 0
-        self._sample_done = asyncio.Event()
-
-    def sample_done_callback(self) -> None:
-        self.samples_in_flight -= 1
-        self._sample_done.set()
-
-    def has_capacity(self, *, pending_groups: int, group_budget: int) -> bool:
-        """A False return arms the sample wakeup: the event is cleared at the moment
-        we decide to wait, so only later completions wake ``wait_for_progress``."""
-        if pending_groups == 0:
-            # No group in flight means no sample is either; drop orphaned credits
-            # from groups that returned without ever spawning their sample tasks.
-            self.samples_in_flight = 0
-        if self.samples_in_flight + self.group_size <= group_budget * self.group_size:
-            return True
-        self._sample_done.clear()
-        return False
-
-    def on_submit(self, groups: list[list[Sample]]) -> None:
-        self.samples_in_flight += sum(len(group) for group in groups)
-
-    async def wait_for_progress(self, pendings: set[asyncio.Task]) -> tuple[set[asyncio.Task], set[asyncio.Task]]:
-        waiter = asyncio.create_task(self._sample_done.wait())
-        try:
-            done, pending = await asyncio.wait(pendings | {waiter}, return_when=asyncio.FIRST_COMPLETED)
-        finally:
-            waiter.cancel()
-        return done - {waiter}, pending - {waiter}
-
-
-def make_submission_scheduler(args: Namespace, *, default: str) -> GroupLevelSubmission | SampleBackfillSubmission:
-    granularity = args.rollout_submission_granularity or default
-    if granularity == "group":
-        return GroupLevelSubmission()
-    assert granularity == "sample", f"unknown submission granularity: {granularity}"
-    return SampleBackfillSubmission(args.n_samples_per_prompt)
-
-
 async def generate_and_rm_group(
     state: GenerateState,
     group: list[Sample],
@@ -206,9 +146,7 @@ async def generate_and_rm_group(
             current_sampling_params["sampling_seed"] = args.rollout_seed + idx
         task = asyncio.create_task(generate_and_rm(state, sample, current_sampling_params, evaluation=evaluation))
         if sample_done_callback is not None:
-            # Fires for every submitted sample whatever its outcome (success, exception
-            # or cancellation): each sample frees exactly the slot it took, so in-flight
-            # accounting is conserved and concurrency cannot decay over time.
+            # fires on success, exception, and cancellation, so in-flight accounting is conserved
             task.add_done_callback(lambda _task: sample_done_callback())
         tasks.append(task)
